Extract SNVs, CNVs and clones on genes of interest from superFreq output

This function imports superFreq's SNVs, CNAs and clones for one patient and outputs them into a tidy format where every row is a variant.

extract_goi_supefreq(superFreq_R_path = superFreq_R_path,
  superFreq_meta_path = superFreq_meta_path, studyGenes,
  patientID = "D1", ref_genome = "hg38", VAFcut = 0.15,
  time_order = c("Screen", "Cyc1", "Cyc2", "Cyc3", "Cyc4", "Cyc9"))

Arguments

superFreq_R_path
Path to superFreq R folder.

superFreq_meta_path
Path to superFreq run cohort metadata (file ending with .tsv). If the sample names in the NAME field of the metadata file start with a number, you need to add an "X" before the name with paste0() since this is done by default by superFReq.

studyGenes
character vector containing genes of interest.

patientID
a character vector specifying the patient/s id/s for which variants have to be imported.

ref_genome
character vector for the reference genome used in the analysis ('hg38' or 'hg19')

VAFcut
Minimum VAF for a variant found in a patient.

time_order
vector specifying the order of time points, e.g. the levels of the `Time` variable as specified in the TIMEPOINT field in superFReq metadata.